pyproject.toml
[project]
name = "nameu"
version = "0.1.2"
description = "名称处理工具集"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    {name = "HibernalGlow", email = "your.email@example.com"}
]
license = {text = "MIT"}
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "loguru>=0.7.0",
    "send2trash>=1.8.0",
    "TextualLog @ git+https://github.com/HibernalGlow/TextualLog.git",
    "PresetUI @ git+https://github.com/HibernalGlow/PresetUI.git",
    "rich>=13.0.0",
    "typer>=0.9.0",
    "orjson",
    "pangu",
    "pywin32",
    "nanoid",
    # IDSet依赖
    "sqlmodel>=0.0.8",
    "pypinyin>=0.54.0",
    "toml>=0.10.2",
    "tqdm>=4.67.1",
    "charset-normalizer>=3.4.2",
    "PyYAML>=6.0",
    "pyperclip>=1.8.0",
    "pytest>=8.3.5",
    "tomli>=2.2.1",
    "openpyxl>=3.1.5",
    "invoke>=2.2.0",
    "peewee>=3.18.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/YourUsername/nameu"
Issues = "https://github.com/YourUsername/nameu/issues"

[project.scripts]
nameu = "nameu.__main__:main"
namer = "nameu.utils.namer:main"
coveru = "coveru.__main__:main"
idu = "idu.__main__:main"
timeu = "timeu.__main__:main"
nameset = "nameset.__main__:main"
[build-system]
requires = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["*"]
exclude = ["tests*"]

[tool.black]
line-length = 88
target-version = ["py313"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
